Rollback crypto to common
