+++ crypto-android/src/main/java/care/data4life/crypto/Json.kt
@@ -17,17 +17,7 @@
 package care.data4life.crypto
 
 import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
 
-actual enum class KeyVersion constructor(
-    value: String,
-    val symmetricKeySize: Int,
-    val asymmetricKeySize: Int
-) {
-    @field:Json(name = "0")
-    VERSION_0("0", 256, 2048),
-
-    @field:Json(name = "1")
-    VERSION_1("1", 256, 2048);
-
-    val value: Int = value.toInt()
-}
+actual typealias Json = Json
+actual typealias JsonClass = JsonClass
