+++ crypto-common/src/main/kotlin/care/data4life/crypto/ExchangeKey.kt
@@ -16,4 +16,25 @@
 
 package care.data4life.crypto
 
-expect class ExchangeKey
+@JsonClass(generateAdapter = true)
+data class ExchangeKey constructor(
+    @field:Json(name = "t") val type: KeyType,
+    @field:Json(name = "priv") val privateKey: String?,
+    @field:Json(name = "pub") val publicKey: String?,
+    @field:Json(name = "sym") val symmetricKey: String?,
+    @field:Json(name = "v") internal val version: Int?
+) {
+    constructor(
+        type: KeyType,
+        privateKey: String?,
+        publicKey: String?,
+        symmetricKey: String?,
+        version: KeyVersion
+    ) : this(type, privateKey, publicKey, symmetricKey, version.value)
+
+    fun getVersion(): KeyVersion = when (version) {
+        0 -> KeyVersion.VERSION_0
+        1 -> KeyVersion.VERSION_1
+        else -> KeyVersion.VERSION_1
+    }
+}
