+++ crypto-common/src/main/kotlin/care/data4life/crypto/ExchangeKeyFactory.kt
@@ -16,4 +16,64 @@
 
 package care.data4life.crypto
 
-expect object ExchangeKeyFactory
+import care.data4life.crypto.error.CryptoException
+
+object ExchangeKeyFactory {
+
+    fun createKey(keyVersion: KeyVersion, type: KeyType, encryptedKeyBase64: String): ExchangeKey {
+        if (keyVersion !== KeyVersion.VERSION_1) {
+            throw CryptoException.InvalidKeyVersion(keyVersion.value)
+        }
+
+        return when (type) {
+            KeyType.APP_PUBLIC_KEY -> createPublicAppKey(encryptedKeyBase64)
+            KeyType.COMMON_KEY -> createCommonKey(encryptedKeyBase64)
+            KeyType.DATA_KEY -> createDataKey(encryptedKeyBase64)
+            KeyType.ATTACHMENT_KEY -> createAttachmentKey(encryptedKeyBase64)
+            KeyType.TAG_KEY -> createTagKey(encryptedKeyBase64)
+
+            else -> throw CryptoException.InvalidKeyType(type.name)
+        }
+    }
+
+    private fun createPublicAppKey(encryptedAppPublicKeyBase64: String): ExchangeKey {
+        return ExchangeKey(
+            KeyType.APP_PUBLIC_KEY,
+            null,
+            encryptedAppPublicKeyBase64, null,
+            KeyVersion.VERSION_1
+        )
+    }
+
+    private fun createCommonKey(encryptedCommonKeyBase64: String): ExchangeKey {
+        return ExchangeKey(
+            KeyType.COMMON_KEY, null, null,
+            encryptedCommonKeyBase64,
+            KeyVersion.VERSION_1
+        )
+    }
+
+    private fun createDataKey(encryptedDataKeyBase64: String): ExchangeKey {
+        return ExchangeKey(
+            KeyType.DATA_KEY, null, null,
+            encryptedDataKeyBase64,
+            KeyVersion.VERSION_1
+        )
+    }
+
+    private fun createAttachmentKey(encryptedAttachmentKeyBase64: String): ExchangeKey {
+        return ExchangeKey(
+            KeyType.ATTACHMENT_KEY, null, null,
+            encryptedAttachmentKeyBase64,
+            KeyVersion.VERSION_1
+        )
+    }
+
+    private fun createTagKey(encryptedTagKeyBase64: String): ExchangeKey {
+        return ExchangeKey(
+            KeyType.TAG_KEY, null, null,
+            encryptedTagKeyBase64,
+            KeyVersion.VERSION_1
+        )
+    }
+}
