+++ crypto-common/src/main/kotlin/care/data4life/crypto/GCKey.kt
@@ -16,4 +16,29 @@
 
 package care.data4life.crypto
 
-expect class GCKey
+import care.data4life.crypto.security.SecretKeySpec
+import care.data4life.sdk.util.Base64
+
+@JsonClass(generateAdapter = true)
+data class GCKey(
+    val algorithm: GCAESKeyAlgorithm,
+    internal var symmetricKey: GCSymmetricKey?,
+    val keyVersion: Int
+) {
+
+    @field:Json(name = "key")
+    internal var keyBase64: String? = null
+
+    fun getSymmetricKey(): GCSymmetricKey {
+        return symmetricKey
+            ?: GCSymmetricKey(
+                SecretKeySpec(
+                    Base64.decode(keyBase64!!),
+                    algorithm.transformation
+                )
+            ).also { symmetricKey = it }
+    }
+
+    fun getKeyBase64(): String = keyBase64
+        ?: Base64.encodeToString(symmetricKey!!.value.getEncoded()).also { keyBase64 = it }
+}
